# performance: Use split limit in truncatewords